Starting Price Analysis Starter Code
(YAHOO FIX) This file is meant to be an extension of code from pythonprogramming.net except it fixes the problem of not being able to get data from yahoo anymore. Obtain a free Alpha Vantage API key to access pricing and replace your key in the *get_data_from_alpha_vantage()* function where it reads '(ENTER ALPHA VANTAGE KEY HERE)'.

There is a covariance matrix that is included in here, also from pythonprogramming.net

This code is meant as a starting point for further analysis to be implemented. The main contribution here is fixing the yahoo pricing pull problem and providing a free alternative.